Update run function in run_edit_module.py

+++ Edit_Insert/run_edit_module.py
@@ -6,6 +6,7 @@ import datetime
 from datetime import datetime
 import random
 import time
+import numpy as np
 
 old_print = print
 
@@ -134,26 +135,35 @@ file_name: file name of json file that is output of codex-davinci-002 with natur
 They are multiple outputs for a single input depending upon k
 The file has outputs for multiple prompts
 """
-def run(file_name,out_dir="."):
+def run(file_name,out_dir=".",n_itr):
     with open(file_name,"r") as input_fp:
         data = json.load(input_fp)
 
-        output = {"solutions":[]}
+        # PUT REMAINDER OF FUNCTION IN LOOP...
+        for itr in n_itr:
+            TEMPERATURE = np.random.randint(0, 10)/10
+            N_SOLUTIONS = np.random.randint(1, 100)
 
-        total_output_set = set()
-        for solution in data:
-            outputs = run_edit_multiple_op(solution, EDIT_OPERATIONS)
-            total_output_set.update(outputs)
+            output = {"solutions":[]}
 
-        output["solutions"].extend(total_output_set)
+            total_output_set = set()
+            for solution in data:
+                outputs = run_edit_multiple_op(solution, EDIT_OPERATIONS, TEMPERATURE, N_SOLUTIONS)
+                total_output_set.update(outputs)
 
-        # json_output = json.dumps(output)
+            output["solutions"].extend(total_output_set)
+
+            # json_output = json.dumps(output)
 
 
-        with open(f'{out_dir}/codex_edit_solutions.json', 'w') as outfile:
-            json.dump(output["solutions"], outfile)
-        
-        save_strings_to_py_file(output["solutions"], f"{out_dir}/edit_sol_pys")
+            pref = 'edit_' + str(TEMPERATURE) + 'T_' + str(N_SOLUTIONS) + 'k_init_'
+            if not os.path.exists(f"{out_dir}/{pref}{file_name}"):
+                return
+                with open(f'{out_dir}/{pref}{file_name}', 'w') as outfile:
+                    json.dump(output["solutions"], outfile)
+            
+                fold_name = pref + file_name[:-4] + "_pys"
+                save_strings_to_py_file(output["solutions"], f"{out_dir}/{fold_name}")
 
 
 
@@ -161,9 +171,10 @@ def run(file_name,out_dir="."):
 
 if __name__ == "__main__":
 
-    #python3 run_edit_module.py example_output.json
+    #python3 run_edit_module.py example_output.json 100
     input_file_name = sys.argv[1]
     out_dir = os.path.dirname(input_file_name)
+    n_itr = sys.argv[2]
 
     #The following code is to save the example prompt and outputs
     """
@@ -181,7 +192,7 @@ if __name__ == "__main__":
         
 
     sys.stdout = open(f'{out_dir}/edit_out.log', 'w')
-    run(input_file_name,out_dir=out_dir)
+    run(input_file_name,out_dir=out_dir,n_itr)
     sys.stdout.close()
